refactor(lang_model): move stream framing into its own module
With Bedrock there are two ways a streamed body is cut into events, SSE and
the AWS binary event stream, but the SSE helpers lived as private functions in
`rt.rs` while the event-stream decoder sat under `impl/`. Gather both, and the
`Framing` selector, under `impl/framing/` so the runtime file is only the
runtime again. Move-only; no behaviour change.
