+++ src/lang_model/impl/framing/eventstream.rs
