+++ src/lang_model/impl/mod.rs
@@ -1,3 +1,3 @@
 pub(crate) mod api;
-pub(crate) mod eventstream;
+pub(crate) mod framing;
 pub(crate) mod response_format;
